@@ -1,9 +1,8 @@
 module github.com/odpf/shield
 
-go 1.16
+go 1.18
 
 require (
-	cloud.google.com/go/storage v1.18.2 // indirect
 	github.com/MakeNowJust/heredoc v1.0.0
 	github.com/abbot/go-http-auth v0.4.0
 	github.com/authzed/authzed-go v0.3.0
