@@ -9,7 +9,7 @@ type Project struct {
 	Name         string
 	Slug         string
 	Organization Organization
-	Metadata     map[string]string
+	Metadata     map[string]any
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
 }
@@ -18,7 +18,7 @@ type Organization struct {
 	Id        string
 	Name      string
 	Slug      string
-	Metadata  map[string]string
+	Metadata  map[string]any
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
@@ -29,7 +29,7 @@ type Group struct {
 	Slug           string
 	Organization   Organization
 	OrganizationId string `json:"orgId"`
-	Metadata       map[string]string
+	Metadata       map[string]any
 	CreatedAt      time.Time
 	UpdatedAt      time.Time
 }
@@ -40,7 +40,7 @@ type Role struct {
 	Types       []string
 	Namespace   Namespace
 	NamespaceId string
-	Metadata    map[string]string
+	Metadata    map[string]any
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
 }
@@ -77,7 +77,7 @@ type User struct {
 	Id        string
 	Name      string
 	Email     string
-	Metadata  map[string]string
+	Metadata  map[string]any
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
